ref: moved basic query conformance tests
