+++ src/tests/conformance/mod.rs
@@ -1,3 +1,4 @@
+mod basic_query;
 mod handshake;
 mod invalid_message;
 mod messages;
